Move Puzzle and Row out to their own files. Refactor the specs a little. Be clear to use numbers, not strings.
